Report the PROJ version using one of the utilities
A possible reason for certain failures reported by CPAN Testers might perhaps be that the PROJ version is too old and doesn't include the headers/functions that the tests expect it to include. But if Alien::proj is installed, it should also come with the PROJ utilities. This test file uses those utilities to determine and report the PROJ version number for debugging.

Another possible reason might perhaps be that PROJ is installed twice and the wrong installation is used? Not sure how to defend against that.