Add distances functions
The geodesic distance functions seem to always return Inf. PROJ has a unit test for proj_lp_dist that expects a non-infinite number and supposedly succeeds, but that test's code also returns Inf for me.

However, this same behaviour is also exhibited when I compile C code directly, so I guess whatever the issue is, it's not one caused by FFI or this Perl module.