Add new PJ_COORD struct member accessors like xyz_x
This change also introduces a stable API for modifying coordinate values.

It might be possible to build even faster read/write accessors using the private FFI::Platypus::Record::_accessor XS API, but it was never designed to allow that so the result would be fairly brittle.

The old xyz->x syntax has been around for a while, so it should remain available for a long time for backwards compatibility.